+++ packages/worphling/src/cli/Cli.ts
@@ -10,25 +10,19 @@ export class Cli {
 
     private detectFlags(): Flags {
         const args = minimist(process.argv.slice(2), {
-            boolean: ["try-exact-length", "with-sorting", "skip-modified-detection", "force-retranslate-all"],
+            boolean: ["try-exact-length", "with-sorting"],
             default: {
                 "try-exact-length": false,
                 "with-sorting": false,
-                "skip-modified-detection": false,
-                "force-retranslate-all": false,
             },
         });
 
         const isTryingExactLengthEnabled = args["try-exact-length"];
         const isSortingEnabled = args["with-sorting"];
-        const isModifiedDetectionSkipped = args["skip-modified-detection"];
-        const isForceRetranslateAllEnabled = args["force-retranslate-all"];
 
         return {
             isTryingExactLengthEnabled,
             isSortingEnabled,
-            isModifiedDetectionSkipped,
-            isForceRetranslateAllEnabled,
         };
     }
 }